Inverted module dependency between root and modules
This commit consists of two changes:

1. Reworked module export syntax to use `export` statement, rather than
   exports being part of the module declaration
2. Inverted the dependency relationship between modules defined in a
   source file and code defined outside the modules

The first change means that `module m {f} (def f -> ...)` has been
changed to `module m (export {f}; def f -> ...)`. For a module that
exports a lot of definitions (say the prelude) I found it clunky to have
the module definition start with a giant export statement before the
opening parenthesis as if the exports are somehow outside the module
definition.

The second change means that any module defined in a source file is
imported so definitions outside of modules in the source file can use
the modules. The way this used to work is that the definitions in the
source file would be imported in to the modules instead.

Several times over I would write tests and do something like the
following:

```
module m (def f -> ...)
def g a -> m/f(a)
```

which then to my surprise wouldn't work. The way things worked before
this commit, `g` would be accessible inside module `m`, but definitions
in `m` wouldn't be imported to the outside source file scope. The fact
that I would get this wrong myself several times over, made me realise
that it was possibly just counterintuitive and maybe needed changing.