Changed clause separator
One of the ways my LALR syntax is ambigious is how a newline separates
both a clause and a statement. From a readability point of view this
isn't a problem. It's clear that if something is defined using e.g.
`def` then it's not a new clause. However the parser only looks one
token ahead and when that token is a newline, it can't differentiate
between clauses and statements.

I've been trying really hard to figure out how to make the parser work
without changing the syntax. The fact that the ambiguity comes only from
the LALR constraint is annoying and I would have liked to work around
it, but I can't seem to find a way to do that.

So with regrets and a feeling of defeat, I've introduced a _mandatory_
comma separating clause statements.