Make sure line separators between statements are right-associative
Having line separators be right associative makes sure that any
statement after a let construct is `seq`ed together underneath the scope
of the let statement. Take the following example:

```
val f -> (let a = 2
          a.print
          a + 4)
```

When line separators are right associative, the function body of `f`
gets parsed to an AST as follows (in simplified representation):

```
{let, {clause, [{pattern, a}], 2},
      {seq, {application, print, [a]},
            {application, +, [a, 4]}}}
```

However if line separators were left associative then we'd get this
instead:

```
{seq, {let, {clause, [{pattern, a}], 2}
            {application, print, [a]}},
      {application, +, [a, 4]}}
```

This ast is not scoped right, since `a` isn't defined in the second
application.