Introduced `whatever` domain keyword
When monomorphizing an exported function in a module, we would need to
call linearize def with a set of arguments. This posed a problem because
running the linearizer with `any` as all arguments would lead to a
`arguments_not_subset` error when linearizing a function using pattern
matching unless the pattern contained a wildcard clause.

The domain `whatever` solves this by being a subset of any value. That
means that when we linearize a function with `whatever` as the input
arguments, every clause is linearized and the full function tree is
generated.

The rules for `whatever` are as follows:

```
diff(whatever, D) -> none
subset(whatever, D) -> true
union(whatever, D) -> D
intersection(whatever, D) -> D
```

This makes `whatever` an indifferent cuisin of `any` in that for
intersections they both behave the same, but where as `any` will
propogate through the code, `whatever` will only propagate up until the
point where any other domain.

I worry that I've created an abomination with this domain keyword. In
particular I worry that due to how `whatever` works in unions, some code
pathways might not be expanded, although I can't think of examples right
now. Maybe because Ash is crying and we both need to go to bed.