Added Monomorphization -- Simple programs compile and run
This commit is (once again) a massive hodgepodge of changes related to
making `kind.erl` run and the tests in this module pass.

With tests passing in `kind.erl`, we're at a stage where very simple
programs can be compiled and run. Codegen has barely been touched
outside of deleting a bunch of obviously useless code, and I've not
tried to run codegen tests or typegen tests at all.

Changes include:
- Removed type and tagged related functions from `expr_gen` and
  `pattern_gen`
- Added support for `keyword` terms in whitelisting to fix an issue
  where we'd check if an application of a keyword would be allowed
- Rewrote kind module to take in source and libraries. In general we
  monomorphize and compile a single source file only but can add set of
  libraries that might be imported.
- Added support for precomputed env for linearization so that we don't
  have to relinearize the same functions over and over
- A few bugfixes here and there