Fixed issue where module keywords wouldn't get imported
When using an imported keyword like `True` in a local function `f`, the
keyword would be tagged as `<module>/f/True` rather than
`kind/prelude/boolean/True` as I would expect.

The issue was that the tagger when tagging keywords would only tag the
local version (e.g. `f/True`) and not the global version (e.g. `True`).
To fix this I've added another check to see if the keyword exists in
scope before testing if the local version of the keyword exists in
scope.