Formalised `defs` and introduce evaluation of linearized function terms
This commit lumps two very disparage changes together. The first is a
formalisation of how `defs` are represented and evaluated. The second is
how linearized code is evaluated. I'll go through each change in tern,
while apologizing profusely for my improper commit-hygiene.

Formalisation of `defs`:
------------------------

If I define `f` and then call it without parenthesis (e.g. `val g = f`) is
`f` evaluated or is it now a local function? And does this depend on whether
`f` expects any arguments and whether `f` is called in a pattern?

With this commit, I've settled on the following approach:

- A `def` defined with arguments called without parenthesis returns
  a function.
- A `def` defined with no arguments called without parenthesis returns
  an expression. This expression is evaluated at each call site.

Internally a `def` with arguments is stored as a definition covering a
function: `{def, ctx, name, {fun, ctx, clauses}}`. In comparison, a
`def` with no arguments is stored as a definition covering an
expression: `{def, ctx, name, expr}`.

This representation means that there's currently no syntactic way to
refer to a function with zero arguments without evaluating it. If we
define `def f -> beam/random/uniform(10)` and declare that `val g = f`,
then `g` will contain a random number. Not a function which takes zero
arguments.

To accomodate this use case I plan to (BUT HAVEN'T YET) introduce a
syntactic element `@` which turns an expression in to a function with
zero arguments. For example:

```
def main -> (val r = @beam/random/uniform(10)
             [r(), r(), r()])
```

This example will return a list of 3 different random numbers. This
means that a `def` with no arguments that we don't intend to evaluate
can just be prefixed with an `@` at the call site.

I'd like to overload the `@` in a function definition to indicate that
any parameter instead of being evaluated is wrapped in a function. This
provides the user with a way to specify that a function argument is
evaluated similar to a call-by-name fashion, except that by turning it
in to a function, it is clear from the context when the argument is
being referenced and when it is being evaluated.

As an added aside, this would allow us to define `if` and `else` as
follows:

```
def if @expr True -> expr(),
       _ False    -> Nil

def else Nil @otherwise -> otherwise(),
         expr _         -> expr

% This would then be used as follows:
a(i).if(i < 10).else(a(10))
```

Evaluation of linearized function terms
---------------------------------------

When we linearize a function, the linearizer would return `{fun, ctx,
F}` where `F` would be a function that we would need to call with an
argument list in order to evaluate the function tree.

The motivation behind this behavior was that because the linearized function
definition was determined by the input arguments, we would only be able
to know the exact function term given these arguments.

However with the introduction of the `whatever` keyword (commit:
babda9e) it is possible to produce a generalised linearized function,
i.e. a function which produces the same output as a function linearized
with respect to more specific arguments.

As a consequence I've changed so `linearize:term` returns a fully
evaluated tree. There's of course cases where the fully evaluated tree
is much more verbose than the tree evaluated to a specific set of
arguments, but I realised as I made this change that quite often a
project will have a single main function with a narrow domain for the
input arguments.

I'm still a bit doubtful that this change is not going to shoot me
massively in the foot at some point, but if it doesn't, then I think
it's a massive step forward for several reasons:

- Modules can trivially be compiled to erlang modules and export
  functions that can be called from beam
- Internals of the linearize module aren't exposed. Instead the function
  return is fully linearized and ready to be compiled.