Added support for tuple syntax and domains: #{a, b}
Tuples are an integral part of erlang and while the same functionality
can be replicated using lists, there's as far as I can see a few good
reasons to have tuple support in Lark:
- Beam interop: Many functions return tuples or accept tuples as input
- Readability: A List in a structure invites to be prepended while a
  tupple is more static. This difference in "feel" can be important for
  how easy it is to understand a piece of code
- Performance: Tuples have different performance characteristics than
  lists. While I've made no attempts at making Lark a fast language, I
  think there's a certain value in not messing with anybody's
  expectation of roughly how expensive something is

In the language, a tuple is syntactically defined as `#{a, b, ...}`
where `a` and `b` can be expression or a pattern of any output domain.
This representation is the same both for patterns and expressions.