Fixed issue where module would import itself
If a source file contained a module `m` and the root module defined an
import of some def in `m`, this import would also be present in `m`, causing the
tagger to error because the def is both part of localscope and imported
scope.

The behavior of root imports showing up in modules defined in the source
file is desired. My thinking is that a source file can specify a few top
level imports that are then present in all modules without causing a
bunch of duplicate code. Any import we want in a specific module can be
done inside this module.

To fix the issue, I've filtered the imports we transfer to any module
from the source file so the module itself is not a prefix.