Simplifying tagged constructors
In e.g. `def t a b -> (T: #(a, b))`, it used to be the case that the tagged
constructor `T(...)` would be created on the basis that any non litteral
in the value that was tagged would be substituted for an argument to the
constructor.

With this commit, I'm simplifying this behaviour with the simple rule
that a tagged constructor has one argument, and that this argument must
have the same domain as the domain following the constructor name. So in
the example above, the argument must be shaped like a tuple with two
inputs.

In the case of variables (like in the example above), they are
substituted for the domain `any` in the tagged constructor.