Deleted type checking artefacts
Given that we no longer have types per se, there's no point in keeping
all the old files and tests... At least not in their current form.

These tests are all failing because they make assumptions about the
compiler that are no longer true. It's a shame, but it's also ok. A lot
has changed since they were written.

I've preserved the tests as comments at the bottom of the end-to-end
tests. There's 67 tests in total on my last count. That's a lot of
tests, and while I think there's value in porting them, I think it would
suck the life out of this project to try to do that in one fell swoop.
For that reason I've decided that I'll try to slowly chip away at them
and improve test coverage over the current version, but I also want to
feel excited about adding new features to the compiler without having to
feel like I first need to do two months worth of work before I can get
to the fun part.