Began suppport for overloading defs
This commit adds support for importing multiple functions with the same
name (e.g. `map` from both `beam/lists` and `beam/maps`) as well as
defining local functions with the same name as an imported function.

When this happens, a new AST element is defined called `overloaded`
which contains a list of the underlying qualified symbols.

This is a step on the path rowards supporting def overloading where we
use the appropriate function based on the domain of the arguments.

To implement this feature, here's the next steps:
 - Add support for `overloaded` ast term in linearize. Linearize with
   respect to all possible functions and see if domain is none. If only
   one function is left, then use it. Otherwise, myabe use a case
   clause.
 - For extensibility, we'd like for a library to add support for using a
   data type that the library doesn't know about. For example you might
   want to call a library with a functor that you've imported yourself.
   In this case we need a way to make the library responsive to the
   caller's import scope. We might just want to pass the import scope
   along to the library call.