Don't propagate import of local modules to avoid cyclical dependencies
Normally any imports in the root module also applies to local modules.
However imports of other local modules should be excluded. Otherwise we
can easily end up in a situation with cyclical dependencies, e.g. by
declaring two modules and importing both of them in the root module.

If we apply these two imports to the two local modules, they'll each
import the other. That's not going to fly.